refactor, tidy, and C901 fixes
