+++ requirements.txt
@@ -19,7 +19,6 @@ stringcase
 tinycss2
 flask
 fonttools
-depq
 trimesh
 scipy==1.7.3
